requirements.txt
flask
requests
psycopg2-binary
google-generativeai
scikit-learn
pillow
tensorflow
numpy
pandas
opencv-python
gunicorn